unstandardize_pressure returns int

+++ data_parsing_helpers/file_helpers.py
@@ -96,7 +96,7 @@ def unbin_data(binned_data: List[int], n_bins: int = 256, n_bits: int = 16) -> L
         return unmud_pressure
 
     def unstandardize_pressure(standardized_pressure: float) -> int:
-        return standardized_pressure * max_pressure
+        return int(standardized_pressure * max_pressure)
 
     unbinned_data = list(map(unbin, binned_data))
     reverse_mu_data = list(map(_reverse_mus_law, unbinned_data))
